[refactor] Token 관련 서비스 인터페이스를 하나로 통합 #14
