--- src/main/kotlin/dev/codeswamp/global/auth/domain/service/TokenStoreService.kt
@@ -1,9 +1,0 @@
-package dev.codeswamp.global.auth.domain.service
-
-import dev.codeswamp.global.auth.domain.model.authToken.ValidatedRefreshToken
-
-interface TokenStoreService {
-    fun storeRefreshToken(token: ValidatedRefreshToken)
-    fun rotateRefreshToken(token: String, newToken: ValidatedRefreshToken)
-    fun findRefreshTokenByUserId(userId: Long): ValidatedRefreshToken?
-}
