--- src/main/kotlin/dev/codeswamp/global/auth/domain/service/TokenValidationService.kt
@@ -1,11 +1,0 @@
-package dev.codeswamp.global.auth.domain.service
-
-import dev.codeswamp.global.auth.domain.model.authToken.RawAccessToken
-import dev.codeswamp.global.auth.domain.model.authToken.ValidatedAccessToken
-import dev.codeswamp.global.auth.domain.model.authToken.RawRefreshToken
-import dev.codeswamp.global.auth.domain.model.authToken.ValidatedRefreshToken
-
-interface TokenValidationService {
-    fun validateAccessToken(accessToken: RawAccessToken) : ValidatedAccessToken
-    fun validateRefreshToken(refreshToken: RawRefreshToken) : ValidatedRefreshToken
-}
